splitting ExtDeepLinkController in two
